♻️ refactor: move §4.3 shadow tokens to their own file
Pure move, values unchanged. `DesignTokens.swift` was at 367 of swiftlint's
400-line `file_length` cap, which the pre-commit hook runs under `--strict`
— not enough headroom for the near-black retint (#1378) and the "why this
tint" rationale it needs. Same reason `DesignTokens+OccluderShadows.swift`
was split out, and the new file keeps the `DesignTokens` prefix for the
same reason too: `scripts/check_design_tokens_css.py` globs
`DesignTokens*.swift` and would silently stop mirroring these values.

367 -> 325 lines.

Part of #1378
